requirements.txt
amqp==5.3.1
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
Django==5.1.3
django-celery-results==2.5.1
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-redis==5.4.0
django-taggit==6.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
l18n==2021.3
laces==0.1.1
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
pillow_heif==0.20.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.2.0
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
telepath==0.3.1
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wagtail==6.3.1
wcwidth==0.2.13
Willow==1.9.0
